Remove Card folder and move Card.jsx to components folder
