+++ src/App.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
-import Card from "./card/Card.jsx";
+import Card from "./components/Card.jsx";
 import Layout from "./components/Layout";
 import { bodies } from "./mockData/mockData";
