--- src/card/styles/Card.module.css
@@ -1,4 +1,0 @@
-.cardName {
-  color:aliceblue;
-}
-
